+++ src/main/java/br/com/spring/ecommerce/service/TokenService.java
@@ -44,9 +44,9 @@ public class TokenService {
 		}
 	}
 
-	public Long getIdUser(String token) {
+	public Integer getIdUser(String token) {
 		Claims claims = Jwts.parser().setSigningKey(this.secret).parseClaimsJws(token).getBody();
-		return Long.parseLong(claims.getSubject());
+		return Integer.parseInt(claims.getSubject());
 	}
 
 }
